Split components in separate files
